adicionando etapa de _adicionar_dados_banco e adicionando validação para valor_compra_produto e quantidade

+++ Controle-Estoque-Loja-Roupa/banco_de_dados_glow_modas.py
@@ -9,7 +9,7 @@ cur.execute("""CREATE TABLE IF NOT EXISTS produtos (
     nome_produto TEXT NOT NULL,
     tipo_produto TEXT NOT NULL,
     valor_compra_produto REAL NOT NULL,
-    quantidade_estoque REAL NOT NULL,
+    quantidade_estoque INTEGER NOT NULL,
     data_compra TEXT NOT NULL
 );""")
 #OBS: TALVEZ quantidade_estoque SEJA UMA NOVA TABELA
